map upsert
Also reshape integration tests a little.

Interestingly I've discovered that generating a lot of code in a test
module can cause that test module to run slowly. I had two `assert`
calls in `Support.Integration.~>/2` and there was a ~0.7s pause when
this test module started to run. Removing one of them resulted in the
tests going from ~1.4s to ~0.7s.